azure_data_etl/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
langchain
langchain-community
langchain-google-genai
langchain-cohere
langchain-pinecone
langchain-mongodb
python-dotenv
torch